src/requirements-gradio.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
amazon-textract-caller==0.2.3
amazon-textract-response-parser==1.0.2
amazon-textract-textractor==1.7.10
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
attrs==23.2.0
aws-cdk-lib==2.140.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
backoff==2.2.1
beautifulsoup4==4.12.3
bert-score==0.3.13
boto3==1.34.99
boto3-stubs==1.34.99
botocore==1.34.99
botocore-stubs==1.34.94
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
constructs==10.3.0
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.5
datasets==2.19.1
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docx2txt==0.8
editdistance==0.8.1
effdet==0.4.1
email_validator==2.1.1
emoji==2.11.1
et-xmlfile==1.1.0
evaluate==0.4.2
fastapi==0.111.0
fastapi-cli==0.0.2
ffmpy==0.3.2
filelock==3.14.0
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-vision==3.7.2
googleapis-common-protos==1.63.0
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsii==1.98.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.51
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.54
layoutparser==0.3.4
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
multiprocess==0.70.16
mypy-boto3-bedrock-agent-runtime==1.34.90
mypy-boto3-bedrock-runtime==1.34.90
mypy-extensions==1.0.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.17.3
opencv-python==4.9.0.80
openpyxl==3.1.2
opensearch-py==2.5.0
ordered-set==4.1.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.15.1
pillow==10.3.0
pillow_heif==0.16.0
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
publication==0.0.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydub==0.25.1
Pygments==2.18.0
pypandoc==1.13
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.29.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.0
referencing==0.35.1
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.3
s3transfer==0.10.1
safetensors==0.4.3
scipy==1.13.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminal==0.4.0
timm==0.9.16
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.2
typeguard==2.13.3
typer==0.12.3
types-awscrt==0.20.9
types-s3transfer==0.10.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
unstructured==0.13.6
unstructured-client==0.22.0
unstructured-inference==0.7.29
unstructured.pytesseract==0.3.12
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.3
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
